pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
dependencies=[
    "datasets",
    "faiss-cpu",
    "huggingface_hub",
    "matplotlib",
    "numpy",
    "pandas",
    "plotly",
    "optuna",
    "seaborn",
    "sentence_transformers", 
    "scikit-learn",
    "tqdm",
    "umap-learn"
]

name = "easy_text_clustering"
version = "1.3.0"
authors = [
  { name="billingsmoore", email="billingsmoore@gmail.com" },
]
description = "User-friendly, low-code text clustering"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[tool.hatch.build.targets.wheel]
packages = ["src/easy_text_clustering"]

[project.urls]
Homepage = "https://github.com/billingsmoore/easy_text_clustering"
Issues = "https://github.com/billingsmoore/easy_text_clustering/issues"